Rename controllers files
