+++ src/server/components/sitemap/controllers.ts
@@ -1,6 +1,22 @@
-import { Express } from "express";
-import { sitemapRouter } from "./router";
+import { Request, Response } from "express";
+import { ServiceType } from "server/models/types";
+import { countriesList } from "server/services/metadata";
+import { listsRoutes, getServiceLabel } from "server/components/lists";
 
-export async function initSitemap(server: Express): Promise<void> {
-  server.use(sitemapRouter);
+export function sitemapController(req: Request, res: Response): void {
+  const sections = Object.keys(ServiceType).map(serviceType => {
+    return {
+      title: `Find ${getServiceLabel(serviceType)} per country`,
+      links: countriesList.map(({ value }) => {
+        return {
+          title: value,
+          href: `${listsRoutes.finder}?country=${value}&serviceType=${serviceType}`,
+        };
+      }),
+    };
+  });
+  
+  res.render("sitemap.html", {
+    sections,
+  });
 }
