+++ src/server/components/sitemap/helpers.ts
@@ -1,0 +1,6 @@
+import { Express } from "express";
+import { sitemapRouter } from "./router";
+
+export async function initSitemap(server: Express): Promise<void> {
+  server.use(sitemapRouter);
+}
