+++ src/server/components/sitemap/router.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { sitemapRoute } from "./routes";
-import { sitemapController } from "./sitemap";
+import { sitemapController } from "./controllers";
 
 export const sitemapRouter = express.Router();
